internal: move ValidationResult to base-core
