+++ generators/base-core/generator.ts
@@ -38,7 +38,7 @@ import { CUSTOM_PRIORITIES, PRIORITY_NAMES, PRIORITY_PREFIX, QUEUES } from '../b
 import { createJHipster7Context, joinCallbacks } from '../base/support/index.js';
 import { CRLF, LF, type Logger, hasCrlr, normalizeLineEndings, removeFieldsWithNullishValues } from '../../lib/utils/index.js';
 
-import type { CascatedEditFileCallback, EditFileCallback, EditFileOptions, ValidationResult, WriteFileOptions } from '../base/api.js';
+import type { CascatedEditFileCallback, EditFileCallback, EditFileOptions, WriteFileOptions } from '../base/api.js';
 import {
   type ExportGeneratorOptionsFromCommand,
   type ExportStoragePropertiesFromCommand,
@@ -59,7 +59,13 @@ import { dockerPlaceholderGenerator } from '../docker/utils.js';
 import { extractArgumentsFromConfigs } from '../../lib/command/index.js';
 import type GeneratorsByNamespace from '../types.js';
 import type { GeneratorBaseCore } from '../index.js';
-import type { Config as CoreConfig, Features as CoreFeatures, Options as CoreOptions, GenericTaskGroup } from '../base-core/types.js';
+import type {
+  Config as CoreConfig,
+  Features as CoreFeatures,
+  Options as CoreOptions,
+  GenericTaskGroup,
+  ValidationResult,
+} from './types.js';
 import { convertWriteFileSectionsToBlocks } from './internal/index.js';
 
 const {
