+++ generators/server/support/database.ts
@@ -20,7 +20,7 @@ import crypto from 'crypto';
 
 import { databaseTypes, fieldTypes } from '../../../lib/jhipster/index.js';
 import { databaseData } from '../../spring-data-relational/support/index.js';
-import type { ValidationResult } from '../../base/api.js';
+import type { ValidationResult } from '../../base-core/index.js';
 import { hibernateSnakeCase } from './string.js';
 
 const dbTypes = fieldTypes;
