+++ generators/server/support/relationship.ts
@@ -18,7 +18,7 @@
  */
 
 import { addOtherRelationship } from '../../base-application/support/index.js';
-import type { ValidationResult } from '../../base/api.js';
+import type { ValidationResult } from '../../base-core/index.js';
 import { databaseTypes } from '../../../lib/jhipster/index.js';
 import type { Entity } from '../../../lib/types/application/entity.js';
 import type { Relationship } from '../../base-application/index.js';
